Improved interpolation call using **kwargs for versatility.
This in this version the interpolation method is called as

def interpolate_sample(self, *argv, sample=None, nodes=None, **kwargs):

where **kwargs contains the parameters necessary to perform the interpolation. As these parameters change according to the chosen method a more versatile method is obtained. The main advantage is that new interpolation methods can be implemented without the need to modify interpolate_sample.